configure: Support STDJARLOC environment variable
The standard location for the installation of jar files varies between
distros. This project defaults to Debian standards. The location needs
to be known during build and install as the Classpath has to be built.
So for distributions which differ from the defaulted Debian standard,
the jar location can be supplied as an environment variable STDJARLOC
which is then used accordingly.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>