bash: Add config file for HiDPI scaling factor
Java Swing  does not cope with HiDPI setups, the application appears so
tiny as to be unusable. An automatic discovery solution to determine if
the environment is HiDPI or not has not been found. Therefore introduce
a config file with a scaling factor parameter. The scaling factor is
zero for non-HiDPI systems and possibly 2.0 for HiDPI systems.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>